requirements.txt
PyGithub
Pyinstaller